improve

+++ vm-guest-agent/az-check-vm-guest-agent-status.ps1
@@ -1,17 +1,29 @@
 # optional parameters
 param (
-    [string]$subscriptionId = "",
+    [string]$subscriptionName = "",
     [string]$resourceGroup = "",
     [string]$vmName = "",
-    [string]$path = './report.csv'
+    [string]$path = './report.csv',
+    [bool]$debug = $FALSE
+
 )
 
 # clean up existing report file
-Write-Host "Remove item at $path ..."
-Remove-Item -Path $path
+
+if (Test-Path $path) { 
+    Write-Host "Remove item at $path ..."
+    Remove-Item -Path $path 
+}
+
+if ($debug) {
+    Write-Host "subscriptionName: " $subscriptionName
+    Write-Host "resource: " $resourceGroup
+    Write-Host "vmName: " $vmName
+    Write-Host "path: " $path
+}
 
 # construct the subscription ids
-$subIds = $subscriptionId -ne "" ? ((Get-AzSubscription -SubscriptionId $subscriptionId | Where-Object {$_.State -ne 'Disabled'}).Id) : ((Get-AzSubscription | Where-Object {$_.State -ne 'Disabled'}).Id)
+$subIds = $subscriptionName -ne "" ? ((Get-AzSubscription -SubscriptionName $subscriptionName | Where-Object {$_.State -ne 'Disabled'}).Id) : ((Get-AzSubscription | Where-Object {$_.State -ne 'Disabled'}).Id)
 
 # loop through all subscriptions 
 $resultsAll = @(foreach ($subId in $subIds) {
@@ -22,13 +34,14 @@ $resultsAll = @(foreach ($subId in $subIds) {
     
     # get all VMs
     $vms = ($resourceGroup -ne "" -and $vmName -ne "") ? (Get-AzVM -ResourceGroupName $resourceGroup -Name $vmName) : (Get-AzVM) 
+    Write-Host "Name `t`t CreateOption `t`t Status  `t`t VMAgentVersion `t`t OsName"
     $results = @(foreach ($vm in $vms) {
         
-        $vmInstanceView = Get-AzVM -Status -ResourceGroup $vm.ResourceGroupName -Name $vm.Name 
+        $vmInstanceView = Get-AzVM -ResourceGroup $vm.ResourceGroupName -Name $vm.Name -Status
         $agent = $vmInstanceView | Select -ExpandProperty VMAgent | Select VMAgentVersion
         $status = $vmInstanceView | Select -ExpandProperty Statuses|  Select-Object -Last 1 # last status object indicate vm status
         
-        Write-Host "Found VM: " $vm.Name
+        Write-Host $vm.Name "`t`t" $vm.StorageProfile.OsDisk.CreateOption "`t`t" $status.Code  "`t`t" $agent.VMAgentVersion "`t`t" $vmInstanceView.OsName
 
         # construct the custom object
         [PSCustomObject]@{
@@ -37,6 +50,7 @@ $resultsAll = @(foreach ($subId in $subIds) {
             vmName = $vm.Name
             rgName = $vm.ResourceGroupName
             location = $vm.Location
+            vmOsName = $vmInstanceView.OsName
             vmOsType = $vm.StorageProfile.OsDisk.OsType
             vmOsDiskCreateOption = $vm.StorageProfile.OsDisk.CreateOption
             vmStatus = $status.Code
